backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
